+++ out/server/server.js
@@ -1134,8 +1134,7 @@ let availableSyntax = {
         lensType: SyntaxLensType.Reference,
         validTokens: [],
         commands: [
-            "aoe2ai.editor.viewConstantUsage",
-            "aoe2ai.editor.viewConstantMisuse"
+            "aoe2ai.editor.viewConstantUsage"
         ],
         dataTemplate: (match) => {
             let data = {};
@@ -1265,13 +1264,6 @@ connection.onCodeLensResolve((lens) => {
                             position: [lens.range.start.line, lens.range.start.character],
                             ranges: findConstantRefs(data.constantName, doc.getText())
                         }]
-                };
-            }
-            else {
-                newLens.command = {
-                    command: availableSyntax.constant.commands[1],
-                    title: `View Instances of ERR2012 (Current Document, ${data.constantName})`,
-                    arguments: [data.constantName]
                 };
             }
             break;